fixed some typing issue & updated order controller (#79)

+++ src/main/java/com/backend/orderhere/dto/OrderDishDTO.java
@@ -2,10 +2,12 @@ package com.backend.orderhere.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class OrderDishDTO {
   private Integer dishId;
   private String dishName;
   private Integer dishQuantity;
-  private Double dishPrice;
+  private BigDecimal dishPrice;
 }
